Move shared duo firmware into shared dir
