+++ shared/duo/MathFunctions.h
