+++ shared/duo/MidiFunctions.h
