+++ shared/duo/Sequencer.h
