+++ shared/duo/effect_custom_envelope.cpp
