+++ shared/duo/envelope.h
