+++ shared/duo/note_stack.h
