+++ shared/duo/seq.cpp
