+++ shared/duo/seq.h
