+++ shared/duo/synth_params.h
